requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-docx==1.0.1
PyPDF2==3.0.1
stripe==7.6.0
openai==1.3.5
alembic==1.12.1